Refactored "activity" modules into "run" modules
