+++ Health/src/main/AndroidManifest.xml
@@ -85,7 +85,7 @@
             android:value="@string/google_maps_key" />
 
         <activity
-            android:name=".activity.ActivityFragment"
+            android:name=".run.RunFragment"
             android:label="@string/title_activity_maps"
             android:exported="true" >
         </activity>
