--- Health/src/main/java/com/sciencesquad/health/activity/ActivityModule.java
@@ -1,5 +1,0 @@
-package com.sciencesquad.health.activity;
-
-public class ActivityModule {
-    // code
-}
