+++ Health/src/main/java/com/sciencesquad/health/core/MainActivity.java
@@ -10,8 +10,8 @@ import android.support.v7.widget.Toolbar;
 import android.view.MenuItem;
 
 import com.sciencesquad.health.R;
-import com.sciencesquad.health.activity.ActivityFragment;
-import com.sciencesquad.health.activity.ActivityLandingFragment;
+import com.sciencesquad.health.run.RunFragment;
+import com.sciencesquad.health.run.RunLandingFragment;
 import com.sciencesquad.health.core.alarm.AlarmFragment;
 import com.sciencesquad.health.nutrition.NutritionFragment;
 import com.sciencesquad.health.sleep.SleepFragment;
@@ -73,8 +73,8 @@ public class MainActivity extends AppCompatActivity
 		if (id == R.id.nav_run) {
 			getFragmentManager()
 					.beginTransaction()
-					.replace(R.id.content, new ActivityLandingFragment(), ActivityLandingFragment.TAG)
-					.addToBackStack(ActivityFragment.TAG)
+					.replace(R.id.content, new RunLandingFragment(), RunLandingFragment.TAG)
+					.addToBackStack(RunFragment.TAG)
 					.commit();
 		} else if (id == R.id.nav_sleep) {
 			getFragmentManager()
