+++ Health/src/main/java/com/sciencesquad/health/run/RunLandingFragment.java
@@ -1,4 +1,4 @@
-package com.sciencesquad.health.activity;
+package com.sciencesquad.health.run;
 
 import android.os.Bundle;
 import android.app.Fragment;
@@ -9,15 +9,15 @@ import android.widget.Button;
 
 import com.sciencesquad.health.R;
 
-public class ActivityLandingFragment extends Fragment{
-    public static final String TAG = ActivityLandingFragment.class.getSimpleName();
+public class RunLandingFragment extends Fragment{
+    public static final String TAG = RunLandingFragment.class.getSimpleName();
 
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceSate) {
 
         // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_activity_landing, container, false);
+        View view = inflater.inflate(R.layout.fragment_run_landing, container, false);
         //Button
         final Button button1 = (Button) view.findViewById(R.id.buttonFreeRun);
         final Button button2 = (Button) view.findViewById(R.id.buttonBuildRoute);
@@ -44,8 +44,8 @@ public class ActivityLandingFragment extends Fragment{
     public void button1Clicked(View view) {
         getFragmentManager()
                 .beginTransaction()
-                .replace(R.id.content, new ActivityFragment(), ActivityFragment.TAG)
-                .addToBackStack(ActivityFragment.TAG)
+                .replace(R.id.content, new RunFragment(), RunFragment.TAG)
+                .addToBackStack(RunFragment.TAG)
                 .commit();
     }
     public void button2Clicked(View view) {
