+++ Health/src/main/res/layout/fragment_run_landing.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.sciencesquad.health.activity.ActivityLandingFragment">
+    tools:context=".run.RunLandingFragment">
 
     <Button
         android:layout_width="wrap_content"
